Add issue-rename-handler tests
