+++ tests/fixtures/payloads/issues.edited.json
@@ -11,21 +11,21 @@
     "number": 35,
     "title": "testing!",
     "user": {
-      "login": "todo-dev-app[bot]",
+      "login": "test-app[bot]",
       "id": 36559360,
       "avatar_url": "https://avatars1.githubusercontent.com/u/10660468?v=4",
       "gravatar_id": "",
-      "url": "https://api.github.com/users/todo-dev-app%5Bbot%5D",
-      "html_url": "https://github.com/apps/todo-dev-app",
-      "followers_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/followers",
-      "following_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/following{/other_user}",
-      "gists_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/gists{/gist_id}",
-      "starred_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/starred{/owner}{/repo}",
-      "subscriptions_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/subscriptions",
-      "organizations_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/orgs",
-      "repos_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/repos",
-      "events_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/events{/privacy}",
-      "received_events_url": "https://api.github.com/users/todo-dev-app%5Bbot%5D/received_events",
+      "url": "https://api.github.com/users/test-app%5Bbot%5D",
+      "html_url": "https://github.com/apps/test-app",
+      "followers_url": "https://api.github.com/users/test-app%5Bbot%5D/followers",
+      "following_url": "https://api.github.com/users/test-app%5Bbot%5D/following{/other_user}",
+      "gists_url": "https://api.github.com/users/test-app%5Bbot%5D/gists{/gist_id}",
+      "starred_url": "https://api.github.com/users/test-app%5Bbot%5D/starred{/owner}{/repo}",
+      "subscriptions_url": "https://api.github.com/users/test-app%5Bbot%5D/subscriptions",
+      "organizations_url": "https://api.github.com/users/test-app%5Bbot%5D/orgs",
+      "repos_url": "https://api.github.com/users/test-app%5Bbot%5D/repos",
+      "events_url": "https://api.github.com/users/test-app%5Bbot%5D/events{/privacy}",
+      "received_events_url": "https://api.github.com/users/test-app%5Bbot%5D/received_events",
       "type": "Bot",
       "site_admin": false
     },
@@ -48,7 +48,7 @@
   },
   "changes": {
     "title": {
-      "from": "sdfasdfasd fsdf sdfsd afsd"
+      "from": "I have been edited"
     }
   },
   "repository": {
